requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=2.0.0
pydantic>=2.7.0
python-dotenv>=1.0.1
browser-use>=0.9.0
playwright>=1.48.0
requests>=2.32.0
